Settings: appwidgetloader...load all providers to do a proper filter
google fixed up the keyguard code in frameworks but did not
adjust in settings the widget filter for it. At least if google
would be consisten they can completly get rid of the filter methods
in settings due that they bypass already the right filter for fetching the
provider now (4.2.2 that was broken). Whatever we do not want
to get rid of it due that we have the show all widgets button
on lockscreen which needs a filtering dependend if the user
pressed the button or not. So fetch all widget providers and
let proper work googles forgoten code

Change-Id: Ib562a485ff6c901888491958ce986fb6372ed70e