Add warning dialog and set root default to disable
due that the daemon need to be started one time to get root which
means on a fresh install su is disabled set the default value in
settings as well to disabled to reflect the current status right

As well added to LCD density changer a warning dialog if the user
does not have root.

This will be revisited on a later point to change it without root
and use instead the new selinux api. But for now it is fine

Change-Id: Ife54f30885c445a4fc9c0cb5da46787fcf9c5e32

Conflicts:
	res/values/slim_strings.xml