Settings: remove always transparency mode navbar/statusbar 1/2
problem is that the transparency on most app totally break due
that a lot apps do not draw behind navbar and statusbar. Nearly
all user never use always mode and just use it on launcher
and lockcreen. So let us remove it

as well a bit cleanup in arrays and strings

Change-Id: Ic6d65d96a6665071689f3bc607c11aeadd29d8b8

Conflicts:
	res/values/liquid_arrays.xml
	res/values/slim_strings.xml
	res/xml/navbar_style_dimen_settings.xml